Reduce maxConnections (#17)
Reduce maxConnections from the default of 20 to 10 as Heroku's Postgres
Hobby Dev tier limits the total number of connections to 20. When using
the max number of connections it is impossible to connect to the db
using external tools or to run migrations. numThreads was also changed
as it must be equal to or less than maxConnections.